refactor(token-guard): dont use an abstract token guard
